move internal logrepl code
